Update index.html
Changed the flag used for Kurmanji (Northern Kurdish) from the Turkish flag to the Kurdish flag (Ala Rengîn), which is formally used by the Kurdistan Regional Government in Iraq. 

While Kurmanji is being spoken by Kurds in Turkey, the Turkish flag doesn't represent the Kurmanji or the Kurdish people. Thus, it's better to use the Kurdish flag that is used for Sorani (Central Kurdish). 

This is a fact and not an opinion!